ANOVA Test for each gene

This function is to test the two-way significance of cell type specific expression cross subjects.

Anova_info(eset, non.zero = TRUE, markers = NULL, clusters, samples,
  select.ct = NULL, num.info = 25, ...)

Arguments

eset
single cell ExpressionSet

non.zero
logical, default as TRUE. If TRUE, we only use gene that have non-zero expression

markers
vector of characters, default as NULL. If NULL, use all genes in eset

clusters
character, the name of phenoData used as clusters

samples
character, the name of phenoData used as samples

select.ct
vector of cell types included, default as NULL. If NULL, include all cell types in x

num.info
numeric, number of selected gene for each cell type. Default at 25

Value

a list of

F statistics of samples;

F statistics of clusters;

F statistics of two-way ANOVA;

selected informative genes: high F statistics for clusters compare to samples