Estimate cell type proportion with MuSiC and NNLS

Estimate cell type proportion with MuSiC and NNLS

music.basic(Y, X, S, Sigma, iter.max, nu, eps)

Arguments

Y
vector of bulk tissue expression

X
matrix, Signature matrix

S
vector of Avg. Library size

Sigma
matrix of Subject level variation (gene * cell type)

iter.max
numeric, maximum iteration number

nu
regulation parameter, take care of weight when taking recipical

eps
Thredshold of convergence

Value

a list with elements:

p.nnls: vector of cell type proportions estimated by nnls (normalized q.nnls, sum to 1);

q.nnls: vector of original estimation from nnls;

fit.nnls: fitted value of `Y` from nnls;

resid.nnls: vector of residual value from nnls;

p.weight: vector of cell type proportions estimated by weighted-nnls (normalized q.weight, sum to 1);

q.weight: vector of original estimation from weight-nnls;

fit.weight: fitted value of `Y` from weighted-nnls;

resid.weight: vector of residual value from weighted-nnls;

weight.gene: vector of weight calculated from weighted-nnls;

converge: character in form of 'Reach Maxiter' or 'Converge at #m';

rsd: vector of residual value from weighted-nnls transfromed by weight;

R.squared: vector of R square of weighted-nnls;

var.p: vector of variance of weighted-nnls estimated cell type proportions.