Scaling bulk data and signature matrix and estimate cell type proportion

Scaling bulk data and signature matrix and estimate cell type proportion

music.iter(Y, D, S, Sigma, iter.max = 1000, nu = 1e-04, eps = 0.01,
  inter.as.bseq = F, normalize = F)

Arguments

Y
vector of bulk tissue expression

D
matrix, Signature matrix

S
vector of Avg. Library size

Sigma
matrix of subject level variation (gene * cell type)

iter.max
numeric, maximum iteration number

nu
regulation parameter, when take recipical

eps
thredshold of converagence

inter.as.bseq
logic, substract avg of Y and D as BSEQ-sc did

normalize
logic, divide Y and D by their standard deviation

Value

a list same as nnls.weight.basic