Scaling bulk data and signature matrix and estimate cell type proportion

Scaling bulk data and signature matrix and estimate cell type proportion

music.iter.ct(
  Y,
  D,
  S,
  Sigma.ct,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE
)

Arguments

Y

vector of bulk tissue expression

S

vector of Avg. Library size

Sigma.ct

matrix of Subject level variation with cross-cell-type covariance.

iter.max

numeric, maximum iteration number. Default is 1000

nu

regulation parameter, take care of weight when taking recipical by default nu = 0.0001.

eps

Threshold of convergence. Default is 0.01.

centered

logic, subtract avg of Y and D. Default is FALSE.

normalize

logic, divide Y and D by their standard deviation. Default is FALSE

X

matrix, Signature matrix

Value

a list same as nnls.weight.basic