Cross-subject Corvriance of Relative Abudance

This function is for calculating the cross-subject covariance of relative abundance for selected cell types.

music_Sigma.ct(x, non.zero, markers, clusters, samples, select.ct)

Arguments

x
ExpressionSet, single cell dataset

non.zero
logical, if true, remove all gene with zero expression

markers
vector or list of gene names

clusters
character, the phenoData used as clusters

samples
character,the phenoData used as samples

select.ct
vector of cell types included, default as NULL. If NULL, include all cell types in x

Value

celltype^2 by gene matrix of covariance