MuSiC Deconvolution with Clusters

This function is to calculate the MuSiC deconvolution proportions with clusters

music_prop.cluster(
  bulk.eset,
  sc.eset,
  group.markers,
  groups,
  clusters,
  samples,
  clusters.type,
  verbose = TRUE,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  ...
)

Arguments

bulk.eset
ExpressionSet for bulk data

sc.eset
ExpressionSet for single cell data

group.markers
list of gene names. The list include differential expressed genes within groups. List name must be the same as `clusters.type`.

groups
character, the phenoData of single cell data used as groups;

clusters
character, the phenoData of single cell dataset used as clusters;

samples
character,the phenoData of single cell dataset used as samples;

clusters.type
list of cell types. The list identify groups of similar cell types.

verbose
logical, default as TRUE.

iter.max
numeric, maximum iteration number

nu
regulation parameter, take care of weight when taking recipical

eps
Thredshold of convergence

centered
logic, substract avg of Y and D

normalize
logic, divide Y and D by their standard deviation

Value

matrix of estimated proportions by MuSiC with cluster information.

See also