Set the period to a new value

Change the period of an index. The index must be ordered as fifths (as specified jointly by period and unit).

If the new period is lower, there is a possibility of colliding values. By default they will not be merged and the operation will abort. In such a case, this usually means the data (notes and/or roles) need some attention. Merging can be asked but will strip out any other column.

If columns note and/or role are present, these richer values will be inherited by the new index. This is a chosen feature that allows subtle roles definitions even when switching to 12-EDO. This can be overwritten by calling add_notes() or add_roles() again: both will provide new labels.

Usage

set_period(x, ...)

# S3 method for m_index
set_period(x, to, merge = FALSE, ...)

# S3 method for m_xtibble
set_period(x, to, merge = FALSE, ...)

Arguments

x

An object.

...

Further arguments passed to or from other methods.

to

An integer. The period to set the index to (31, 19 or 12).

merge

A logical. Default is FALSE.

Value

A modified object.

Examples

index(0:1, period = 31L, unit = 18L) |> set_period(to = 12L)
#> <index<7/12>[2]>
#> [1] ~ 0   1
start_with_some_roles(c("P1", "d7", "A5")) |> set_period(to = 12L)
#> # A tibble: 3 × 2
#>         idx role 
#>   <i<7/12>> <chr>
#> 1        -4 A5   
#> 2       ~ 0 P1   
#> 3         3 d7   
start_with_some_roles(c("P1", "d7", "M6")) |> set_period(to = 12, merge = TRUE)
#> ! Duplicates: -9 and 3
#> # A tibble: 2 × 1
#>         idx
#>   <i<7/12>>
#> 1         3
#> 2       ~ 0