Augment or diminish roles

Augment or diminish roles. This will drop any other column. Index will be reset with the richest attributes.

Usage

shift_role(x, roles, delta)

augment(x, roles)

# S3 method for m_xtibble
augment(x, roles)

diminish(x, roles)

# S3 method for m_xtibble
diminish(x, roles)

Arguments

x

An xtibble() object with a column 'role'.

roles

A string, scalar or vector. The roles to augment or diminish.

delta

An integer. The shift. -7L (diminish) or 7L (augment).

Value

A new xtibble() with 2 columns:

idx the modified index with attributes set to period 31 and unit 18,

role the modified roles.

Examples

start_with_mode(degree = 6) |> add_roles() |> augment("m3")
#> Change of attributes
#> ! Attributes of the index were augmented to account for possible richer
#>   context.
#> ℹ period is now 31 and unit 18.
#> 
#> # A tibble: 7 × 2
#>          idx role 
#>   <i<18/31>> <chr>
#> 1         -4 m6   
#> 2         -2 m7   
#> 3         -1 P4   
#> 4        ~ 0 P1   
#> 5          1 P5   
#> 6          2 M2   
#> 7          4 M3   
start_with_mode(degree = 1) |> add_roles() |> diminish("M7")
#> Change of attributes
#> ! Attributes of the index were augmented to account for possible richer
#>   context.
#> ℹ period is now 31 and unit 18.
#> 
#> # A tibble: 7 × 2
#>          idx role 
#>   <i<18/31>> <chr>
#> 1         -2 m7   
#> 2         -1 P4   
#> 3        ~ 0 P1   
#> 4          1 P5   
#> 5          2 M2   
#> 6          3 M6   
#> 7          4 M3