Estimate cell type proportion with MuSiC and NNLS

Estimate cell type proportion with MuSiC and NNLS

music.basic(Y, X, S, Sigma, iter.max, nu, eps)

Arguments

Y

vector of bulk tissue expression

X

matrix, Signature matrix

S

vector of Avg. Library size

Sigma

matrix of Subject level variation (gene * cell type)

iter.max

numeric, maximum iteration number

nu

regulation parameter, take care of weight when taking reciprocal

eps

Threshold of convergence

Value

a list with elements:

p.nnls: vector of cell type proportions estimated by nnls (add up to 1);

q.nnls: vector of original estimation from nnls;

fit.nnls: fitted value of Y from nnls;

resid.nnls: residual value from nnls;

p.weight: vector of cell type proportions estimated by weighted-nnls (add up to 1);

q.weight: vector of original estimation from weight-nnls;

fit.weight: fitted value of Y from weighted-nnls;

resid.weight: residual value from weighted-nnls;

weight.gene: weight calculated from weighted-nnls;

converge: 'Reach Maxiter', 'Converge at m';

rsd: residual value from weighted-nnls transformed by weight;

R.squared: R square of weighted-nnls;

var.p: variance of weighted-nnls estimated cell type proportions;